refactory: change arquiteture
