+++ src/components/body/body.sass
@@ -1,4 +1,4 @@
-@import '../../variables/colors'
+@import '../style/variables/colors'
 
 .body
     max-width: 100%
