+++ src/components/body/cards/cards.sass
@@ -1,5 +1,5 @@
-@import "../../../variables/colors"
-@import '../../../variables/fonts'
+@import "../../style/variables/colors"
+@import '../../style/variables/fonts'
 
 .cards
     width: 90vh
