+++ src/components/body/summary/summary.sass
@@ -1,6 +1,6 @@
-@import "../../../variables/colors"
-@import '../../../variables/fonts'
-@import '../../../variables/keyframes'
+@import "../../style/variables/colors"
+@import '../../style/variables/fonts'
+@import '../../style/variables/keyframes'
 
 .summary
     width: 80%
