+++ src/components/footer/footer.sass
@@ -1,5 +1,5 @@
-@import '../../variables/colors'
-@import '../../variables/fonts'
+@import '../style/variables/colors'
+@import '../style/variables/fonts'
 
 .footerBox
     width: 100%
